+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## DEV
 
+* support endpoint-override map [#59](https://github.com/cognitect-labs/aws-api/issues/59), [#61](https://github.com/cognitect-labs/aws-api/issues/61), [#64](https://github.com/cognitect-labs/aws-api/issues/64)
 * only parse json response body when output-shape is specified [#66](https://github.com/cognitect-labs/aws-api/issues/66)
 
 ## 0.8.273 / 2019-03-01